Change the base framework down one level.
Move attributes class to a separate file
Move from the 4.5 to the 4.0 framework
Obsolete some of the AddAttribute functions to match what we do in JAVA
