+++ COSE/Message.cs
@@ -340,101 +340,7 @@ namespace COSE
 
 
 
-    public class Attributes
-    {
-        protected CBORObject objProtected = CBORObject.NewMap();
-        protected CBORObject objUnprotected = CBORObject.NewMap();
-        protected CBORObject objDontSend = CBORObject.NewMap();
-        protected byte[] externalData = new byte[0];
-        protected byte[] rgbProtected;
-
-        public void AddAttribute(string name, string value, bool fProtected)
-        {
-            if (fProtected) AddProtected(name, value);
-            else AddUnprotected(name, value);
-        }
-
-        public void AddAttribute(string name, CBORObject value, bool fProtected)
-        {
-            if (fProtected) AddProtected(name, value);
-            else AddUnprotected(name, value);
-        }
-
-        public void AddAttribute(CBORObject key, CBORObject value, bool fProtected)
-        {
-            if (fProtected) AddProtected(key, value);
-            else AddUnprotected(key, value);
-        }
-
-        public void AddProtected(string label, string value)
-        {
-            AddProtected(label, CBORObject.FromObject(value)); 
-        }
-
-        public void AddProtected(string label, CBORObject value)
-        {
-            AddProtected(CBORObject.FromObject(label), value);
-        }
-
-        public void AddUnprotected(string label, string value)
-        {
-            AddUnprotected(label, CBORObject.FromObject(label));
-        }
-
-        public void AddUnprotected(string label, CBORObject value)
-        {
-            AddUnprotected(CBORObject.FromObject(label), value);
-        }
-
-        public void AddProtected(CBORObject label, CBORObject value)
-        {
-            RemoveAttribute(label);
-            objProtected.Add(label, value);
-        }
-
-        public void AddUnprotected(CBORObject label, CBORObject value)
-        {
-            RemoveAttribute(label);
-            objUnprotected.Add(label, value);
-        }
-
-        public void AddDontSend(CBORObject label, CBORObject value)
-        {
-            RemoveAttribute(label);
-            objDontSend.Add(label, value);
-        }
-
-        public CBORObject FindAttribute(CBORObject label)
-        {
-            if (objProtected.ContainsKey(label)) return objProtected[label];
-            if (objUnprotected.ContainsKey(label)) return objUnprotected[label];
-            if (objDontSend.ContainsKey(label)) return objDontSend[label];
-            return null;
-        }
-
-        public CBORObject FindAttribute(int label)
-        {
-            return FindAttribute(CBORObject.FromObject(label));
-        }
-
-        public CBORObject FindAttribute(string label)
-        {
-            return FindAttribute(CBORObject.FromObject(label));
-        }
-
-        private void RemoveAttribute(CBORObject label)
-        {
-            if (objProtected.ContainsKey(label)) objProtected.Remove(label);
-            if (objUnprotected.ContainsKey(label)) objUnprotected.Remove(label);
-            if (objDontSend.ContainsKey(label)) objDontSend.Remove(label);
-        }
-
-        public void SetExternalData(byte[] newData)
-        {
-            externalData = newData;
-        }
-    }
-
+ 
         [Serializable]
     public class CoseException : Exception
