+++ COSE/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="BouncyCastle" version="1.8.1" targetFramework="net45" />
-  <package id="PeterO.Cbor" version="2.4.2" targetFramework="net45" />
-  <package id="PeterO.Numbers" version="0.2.2" targetFramework="net45" />
+  <package id="BouncyCastle" version="1.8.1" targetFramework="net40" />
+  <package id="PeterO.Cbor" version="2.4.2" targetFramework="net40" />
+  <package id="PeterO.Numbers" version="0.2.2" targetFramework="net40" />
 </packages>
